refactor: move code to external fils
