+++ src/lib/constants/fixtures/billabong-models.txt
@@ -1,0 +1,54 @@
+ABJKT00515
+ABJKT00572
+ABYWT00268
+ABYZT02314
+ABYZT02314
+ABYZT02363
+ABYZT02364
+ABYZT02365
+ABYZT02366
+ABYZT02367
+ABYZT02368
+ABYZT02369
+ABYZT02370
+ABYZT02372
+ABYZT02373
+ABYZT02374
+ABYZT02401
+ABYZT02514
+ABYZT02514
+ABYZT02515
+ABYZT02518
+UBYZT00516
+ABGNS00106
+ABGNS00106
+ABYTK03005
+ABYTK03006
+ABYWS00224
+ABYWS00225
+G203JMAD
+ABYFT00474
+ABYSF00122
+ABYSF00123
+ABJX300981
+ABJX300994
+ABJX400214
+ABJX400307
+ABJX400507
+ABJX400901
+ABJX400981
+ABJX401018
+ABYBS00450
+ABYBS00453
+ABYBS00454
+ABYBS00455
+ABYBS00459
+ABYBS00464
+ABYBS00468
+ABYBS00485
+ABYBS00488
+ABYBS00489
+ABYBS00496
+ABYBS00500
+ABYJV00135
+ABYJV00135
