+++ src/lib/constants/fixtures/rvca-models.txt
@@ -1,0 +1,262 @@
+AVBKT00162
+AVBNP00106
+AVBWS00130
+AVBWT00168
+AVBWT00170
+AVBWT00171
+AVBWT00172
+AVBWT00173
+AVBWT00177
+AVBWT00178
+AVBZT00147
+AVBZT00476
+AVBZT00497
+AVBZT00501
+AVBZT00507
+AVBZT00518
+AVJBP00109
+AVJHA00173
+AVJHA00199
+AVJHA00202
+AVJHA00205
+AVJHA00208
+AVJKD00173
+AVJKT00275
+AVJKT00277
+AVJKT00283
+AVJKT00287
+AVJKT00297
+AVJNP00131
+AVJNP00239
+AVJNP00273
+AVJNP00285
+AVJNP00291
+AVJNP00294
+AVJNS00198
+AVJNS00213
+AVJSW00173
+AVJSW00191
+AVJSW00199
+AVJSW00200
+AVJSW00201
+AVJSW00202
+AVJWD00220
+AVJWD00241
+AVJWD00309
+AVJWD00311
+AVJWD00312
+AVJWO03001
+AVJWO03004
+AVJWO03005
+AVJWO03006
+AVJWT00194
+AVJWT00217
+AVJWT00218
+AVJWT00225
+AVJWT00226
+AVJWT00227
+AVJWT00228
+AVJX300199
+AVJX300460
+AVJX400130
+AVJX400247
+AVJX400268
+AVJX400423
+AVJX600135
+AVJX600136
+AVJX600137
+AVJX600145
+AVJZT00916
+AVJZT00931
+AVJZT00938
+AVJZT00941
+AVJZT00943
+AVYBA00132
+AVYBP00138
+AVYBS00293
+AVYBS00297
+AVYBS00324
+AVYBS00338
+AVYBS00343
+AVYBS00346
+AVYBS00349
+AVYBS00362
+AVYBT00131
+AVYDP00128
+AVYFT00351
+AVYFT00352
+AVYHA00202
+AVYHA00270
+AVYHA00340
+AVYHA00387
+AVYHA00454
+AVYHA00457
+AVYHA00460
+AVYHA00465
+AVYHA00466
+AVYHA00467
+AVYHA00478
+AVYHA00479
+AVYHA00512
+AVYHA00514
+AVYHA00530
+AVYHA00561
+AVYHA00596
+AVYHA00629
+AVYHA00631
+AVYHA00638
+AVYHA00657
+AVYHA00658
+AVYHA00660
+AVYHA00661
+AVYHA00662
+AVYHA00663
+AVYHA00664
+AVYHA00665
+AVYHA00675
+AVYHA00695
+AVYHY03000
+AVYJK00123
+AVYJK00208
+AVYJK00231
+AVYKT00262
+AVYKT00268
+AVYKT00271
+AVYKT00332
+AVYKT00345
+AVYKT00348
+AVYNP00109
+AVYNP00220
+AVYSF00305
+AVYSF00306
+AVYSF00307
+AVYSF00308
+AVYWS00135
+AVYWS00229
+AVYWS00233
+AVYWS00235
+AVYWS00239
+AVYWS00242
+AVYWS00250
+AVYWS00257
+AVYWS00298
+AVYWS00299
+AVYWS00314
+AVYWS00322
+AVYWS00323
+AVYWS00324
+AVYWS00330
+AVYWS00331
+AVYWS00332
+AVYWS00338
+AVYWS00339
+AVYWS00344
+AVYWS00346
+AVYWS00349
+AVYWS00350
+AVYWS00351
+AVYWT00403
+AVYWT00421
+AVYWT00449
+AVYWT00453
+AVYWT00474
+AVYWT00476
+AVYWT00489
+AVYWT00490
+AVYWT00494
+AVYWT00495
+AVYWT00496
+AVYWT00498
+AVYWT00499
+AVYWT00500
+AVYWT00501
+AVYWT00503
+AVYWT00507
+AVYWT00511
+AVYWT00513
+AVYWT00515
+AVYWT00516
+AVYWT00517
+AVYWT00518
+AVYWT00519
+AVYWT00520
+AVYWT00521
+AVYWT00522
+AVYWT00523
+AVYWT00524
+AVYWT00527
+AVYWT00528
+AVYWT00530
+AVYWT00533
+AVYWT00544
+AVYZT00194
+AVYZT01365
+AVYZT01370
+AVYZT01536
+AVYZT01830
+AVYZT02057
+AVYZT02085
+AVYZT02117
+AVYZT02118
+AVYZT02120
+AVYZT02121
+AVYZT02122
+AVYZT02123
+AVYZT02125
+AVYZT02126
+AVYZT02127
+AVYZT02128
+AVYZT02129
+AVYZT02130
+AVYZT02131
+AVYZT02133
+AVYZT02134
+AVYZT02136
+AVYZT02137
+AVYZT02139
+AVYZT02140
+AVYZT02141
+AVYZT02142
+AVYZT02143
+AVYZT02144
+AVYZT02146
+AVYZT02148
+AVYZT02151
+AVYZT02152
+AVYZT02153
+AVYZT02154
+AVYZT02155
+AVYZT02156
+AVYZT02157
+AVYZT02163
+AVYZT02164
+AVYZT02165
+AVYZT02166
+AVYZT02167
+AVYZT02172
+AVYZT02174
+AVYZT02176
+AVYZT02177
+AVYZT02178
+AVYZT02179
+AVYZT02214
+AVYZT02215
+AVYZT02217
+AVYZT02328
+AVYZT02369
+AVYZT02370
+AVYZT02371
+AVYZT02376
+AVYZT02377
+M1121RWT
+M1182REE
+M2073REE
+M437VRPT
+M503VRDD
+MAHW2RTR
+MAHWURVT
+UVJJK00138
+UVJKT00182
+UVJKT00191
+UVYZT00652
+V9021RSV
