+++ src/lib/types/scraped-result.type.ts
@@ -1,0 +1,5 @@
+export interface ScrapedResult {
+	model: string;
+	success: boolean;
+	materials?: string;
+}
